Moved the config outside of the start server
